+++ app/src/main/java/com/billdoerr/android/geotracker/utils/Physics.java
@@ -1,0 +1,35 @@
+package com.billdoerr.android.geotracker.utils;
+
+/**
+ * Simple class
+ */
+public class Physics {
+
+    private float distance;
+    private float time;
+    private float velocity;
+
+    public float getDistance() {
+        return distance;
+    }
+
+    public void setDistance(float distance) {
+        this.distance = distance;
+    }
+
+    public float getTime() {
+        return time;
+    }
+
+    public void setTime(float time) {
+        this.time = time;
+    }
+
+    public float getVelocity() {
+        return velocity;
+    }
+
+    public void setVelocity(float velocity) {
+        this.velocity = velocity;
+    }
+}